// src/@types/user-types.ts
import { Request } from 'express';

declare global {
	namespace Express {
		interface Request {
			user: {
				id: any;
				name: string;
				email: string;
				verified: boolean;
				avatar?: string;
				followers: number;
				followings: number;
			};
			token: string;
		}
	}
}

export interface CreateUserRequest extends Request {
	body: {
		name: string;
		email: string;
		password: string;
	};
}
export interface VerifyEmailRequest extends Request {
	body: {
		userId: string;
		token: string;
	};
}